refactor: moving Train to separate directory
