+++ CMakeLists.txt
@@ -97,15 +97,15 @@ set(SOURCES
 "world/Track.cpp"
 "world/Traction.cpp"
 "world/TractionPower.cpp"
-"vehicle/Train.cpp"
-"vehicle/Train_Commands_Auxiliary.cpp"
-"vehicle/Train_Commands_Brakes.cpp"
-"vehicle/Train_Commands_Controls.cpp"
-"vehicle/Train_Commands_Doors.cpp"
-"vehicle/Train_Commands_Electrical.cpp"
-"vehicle/Train_Commands_Lights.cpp"
-"vehicle/Train_Commands_Misc.cpp"
-"vehicle/Train_Commands_Radio.cpp"
+"vehicle/train/Train.cpp"
+"vehicle/train/TrainCommandsAuxiliary.cpp"
+"vehicle/train/TrainCommandsBrakes.cpp"
+"vehicle/train/TrainCommandsControls.cpp"
+"vehicle/train/TrainCommandsDoors.cpp"
+"vehicle/train/TrainCommandsElectrical.cpp"
+"vehicle/train/TrainCommandsLights.cpp"
+"vehicle/train/TrainCommandsMisc.cpp"
+"vehicle/train/TrainCommandsRadio.cpp"
 "world/TrkFoll.cpp"
 "vehicle/AirCoupler.cpp"
 "model/AnimModel.cpp"
