+++ application/application.cpp
@@ -18,7 +18,7 @@ http://mozilla.org/MPL/2.0/.
 #include "utilities/utilities.h"
 #include "simulation/simulation.h"
 #include "simulation/simulationsounds.h"
-#include "vehicle/Train.h"
+#include "../vehicle/train/Train.h"
 #include "utilities/dictionary.h"
 #include "scene/sceneeditor.h"
 #include "rendering/renderer.h"
@@ -26,7 +26,7 @@ http://mozilla.org/MPL/2.0/.
 #include "utilities/Logs.h"
 #include "rendering/screenshot.h"
 #include "utilities/translation.h"
-#include "vehicle/Train.h"
+#include "../vehicle/train/Train.h"
 #include "utilities/Timer.h"
 #include "utilities/dictionary.h"
 #include "version_info.h"
