+++ input/drivermouseinput.cpp
@@ -15,7 +15,7 @@ http://mozilla.org/MPL/2.0/.
 #include "utilities/Globals.h"
 #include "utilities/Timer.h"
 #include "simulation/simulation.h"
-#include "vehicle/Train.h"
+#include "../vehicle/train/Train.h"
 #include "model/AnimModel.h"
 #include "rendering/renderer.h"
 #include "application/uilayer.h"
