+++ vehicle/Driver.cpp
@@ -18,7 +18,7 @@ http://mozilla.org/MPL/2.0/.
 #include "utilities/Globals.h"
 #include "utilities/translation.h"
 #include "utilities/Logs.h"
-#include "vehicle/Train.h"
+#include "train/Train.h"
 #include "world/mtable.h"
 #include "vehicle/DynObj.h"
 #include "world/Event.h"
