+++ widgets/vehicleparams.cpp
@@ -3,7 +3,7 @@
 #include "simulation/simulation.h"
 #include "application/driveruipanels.h"
 #include "vehicle/Driver.h"
-#include "vehicle/Train.h"
+#include "../vehicle/train/Train.h"
 
 ui::vehicleparams_panel::vehicleparams_panel(const std::string &vehicle) : ui_panel(std::string(STR("Vehicle parameters")) + ": " + vehicle, false), m_vehicle_name(vehicle)
 {
